backend/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
SQLAlchemy==2.1.0b2
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
PyJWT==2.8.0
werkzeug==3.0.1
bcrypt==4.1.1
